Story-1: Implemented Mapping, applied groupBy in Spark

+++ src/main/scala/com/atamaniv/model/Coordinates.scala
@@ -1,3 +1,3 @@
 package com.atamaniv.model
 
-case class Coordinates(longitude: Long, latitude: Long)
+case class Coordinates(longitude: Double, latitude: Double)
